+++ android/src/main/java/io/bluedot/BluedotErrorReceiver.java
@@ -1,18 +1,13 @@
 package io.bluedot;
 
+import static io.bluedot.EventUtil.sendEvent;
+
 import android.content.Context;
-import androidx.annotation.Nullable;
 import au.com.bluedot.point.net.engine.BDError;
 import au.com.bluedot.point.net.engine.BluedotServiceReceiver;
-import com.facebook.react.ReactApplication;
-import com.facebook.react.ReactHost;
-import com.facebook.react.ReactInstanceEventListener;
-import com.facebook.react.bridge.ReactContext;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.bridge.WritableNativeMap;
-import com.facebook.react.modules.core.DeviceEventManagerModule;
 import org.jetbrains.annotations.NotNull;
-
 
 public class BluedotErrorReceiver extends BluedotServiceReceiver {
 
@@ -30,30 +25,5 @@ public class BluedotErrorReceiver extends BluedotServiceReceiver {
         WritableMap error = new WritableNativeMap();
         error.putString("error",bdError.getReason());
         sendEvent(context, "onBluedotServiceError",error);
-    }
-
-    public void sendEvent(Context context,
-                          String eventName,
-                          @Nullable WritableMap params) {
-
-        ReactApplication reactApplication = (ReactApplication) context.getApplicationContext();
-        ReactHost reactHost = reactApplication.getReactHost();
-        ReactContext reactContext = reactHost.getCurrentReactContext();
-
-        if (reactContext != null) {
-            reactContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
-                    .emit(eventName, params);
-        } else {
-            reactHost.addReactInstanceEventListener(
-                    new ReactInstanceEventListener() {
-                        @Override
-                        public void onReactContextInitialized(ReactContext context) {
-                            context.getJSModule(
-                                    DeviceEventManagerModule.RCTDeviceEventEmitter.class)
-                                    .emit(eventName, params);
-                            reactHost.removeReactInstanceEventListener(this);
-                        }
-                    });
-        }
     }
 }