+++ ios/BluedotPointSDK.h
@@ -1,5 +1,6 @@
 #import <React/RCTBridgeModule.h>
 #import <React/RCTEventEmitter.h>
+#import <React/RCTLog.h>
 
 @import BDPointSDK;