+++ ReleaseNotes.md
@@ -1,3 +1,3 @@
 # Bluedot React Native Plugin release notes
 
-- Integrated with Bluedot Android SDK v16.1.1 and iOS SDK 16.0.0
+- Integrated with Bluedot Android SDK v17.1.0 and iOS SDK 17.0.0